% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indepvarBF.R
\name{indepvarBF}
\alias{indepvarBF}
\title{Compute a Bayes Factor for the comparison of independent variances}
\usage{
indepvarBF(x, y, rscale = "medium")
}
\arguments{
\item{x}{first continuous variable}

\item{y}{second continuous variable}

\item{rscale}{The prior width, passed to \code{\link[BayesFactor]{oneWayAOV.Fstat}}}
}
\value{
The bayes factor favoring the alternative hypothesis as 
    compared to the null hypothesis
}
\description{
Compute a Bayes Factor for the comparison of independent variances
}
\details{
This function realizes a test of independent variances by converting
the F value from a Levene test to a Bayes Factor using the function 
\code{\link[BayesFactor]{oneWayAOV.Fstat}}.
}
\note{
As stated it the documentation to \code{\link[BayesFactor]{oneWayAOV.Fstat}}, 
this method should only be used with F values obtained from balanced designs
(i.e., the groups should have equal size).
}
\examples{

# True different variances:
indepvarBF(rnorm(100, mean = 100, sd = 15), rnorm(100, mean = 100, sd = 20))
# True same variances: 
indepvarBF(rnorm(100, mean = 100, sd = 15), rnorm(100, mean = 100, sd = 15))

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
