% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varBF-package.R
\docType{package}
\name{varBF-package}
\alias{varBF}
\alias{varBF-package}
\title{\packageTitle{varBF}}
\description{
\packageDescription{varBF}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab varBF\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2019-11-22\cr
Depends: \tab R (>= 3.4.0)\cr
Encoding: \tab UTF-8\cr
License: \tab MIT + file LICENSE\cr
}
}
\author{
\packageAuthor{varBF}

Maintainer: \packageMaintainer{varBF}
}
