% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varBF.R
\name{varBF}
\alias{varBF}
\title{Compute a Bayes Factor for the comparison of dependent variances}
\usage{
varBF(x, y, ...)
}
\arguments{
\item{x}{first continuous variable}

\item{y}{second continuous variable}

\item{...}{further arguments passed to \code{\link[BayesFactor]{correlationBF}}}
}
\value{
The bayes factor
}
\description{
Compute a Bayes Factor for the comparison of dependent variances
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
