% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depvarBF.R
\name{depvarBF}
\alias{depvarBF}
\title{Compute a Bayes Factor for the comparison of dependent variances}
\usage{
depvarBF(x, y, rscale = "medium")
}
\arguments{
\item{x}{first continuous variable}

\item{y}{second continuous variable, sorted in the same order as \code{x}}

\item{rscale}{The prior width, passed to \code{\link[BayesFactor]{correlationBF}}}
}
\value{
The bayes factor
}
\description{
Compute a Bayes Factor for the comparison of dependent variances
}
\details{
This function realizes a test of dependent variances by applying
the Morgan-Pitman test that reduces to a test of the "nullity"
of a correlation between the sum and difference of two continuous 
variables.
}
\examples{

depvarBF(sleep$extra[sleep$group == 1], sleep$extra[sleep$group == 2])
depvarBF(iris$Sepal.Length, iris$Petal.Width)
depvarBF(iris$Sepal.Length, iris$Petal.Width, rscale = "ultrawide")

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
