% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varBF.R
\name{depvarBF}
\alias{depvarBF}
\title{Compute a Bayes Factor for the comparison of dependent variances}
\usage{
depvarBF(x, y, ...)
}
\arguments{
\item{x}{first continuous variable}

\item{y}{second continuous variable}

\item{...}{further arguments passed to \code{\link[BayesFactor]{correlationBF}}}
}
\value{
The bayes factor
}
\description{
Compute a Bayes Factor for the comparison of dependent variances
}
\examples{

depvarBF(sleep$extra[sleep$group == 1], sleep$extra[sleep$group == 2])
depvarBF(iris$Sepal.Length, iris$Petal.Width)

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
